Added Icons for Log Entries
Some icons to be used in the context of logging have been added.

All icons are `10mm` high and wide, but it should be possible to scale
them to smaller units (like `1em`) or larger units. The icons are
available as `svg` vector graphics. Since some browser do not
support `svg`s (or support them in a very odd way), I also add `png`
versions of the icons, which then are `32` pixels high and wide.